+++ chromium-edge-launcher/index.js
@@ -1,4 +1,4 @@
 // This provides legacy support for a previously documented require pattern
-//    const edgeLauncher = require('chromium-edge-launcher/chromium-edge-launcher');
+//    const edgeLauncher = require('chromium-edge-launcher/edge-launcher');
 
 module.exports = require('../');