+++ src/edge-launcher.ts
@@ -8,14 +8,14 @@
 import * as childProcess from 'child_process';
 import * as fs from 'fs';
 import * as net from 'net';
-import * as edgeFinder from './edge-finder';
-import {getRandomPort} from './random-port';
-import {DEFAULT_FLAGS} from './flags';
-import {makeTmpDir, defaults, delay, getPlatform, toWinDirFormat, InvalidUserDataDirectoryError, UnsupportedPlatformError, EdgeNotInstalledError} from './utils';
+import * as edgeFinder from './edge-finder.js';
+import {getRandomPort} from './random-port.js';
+import {DEFAULT_FLAGS} from './flags.js';
+import {makeTmpDir, defaults, delay, getPlatform, toWin32Path, InvalidUserDataDirectoryError, UnsupportedPlatformError, EdgeNotInstalledError} from './utils.js';
 import {ChildProcess} from 'child_process';
-const log = require('lighthouse-logger');
-const spawn = childProcess.spawn;
-const execSync = childProcess.execSync;
+import {spawn, spawnSync} from 'child_process';
+import log from 'lighthouse-logger';
+
 const isWsl = getPlatform() === 'wsl';
 const isWindows = getPlatform() === 'win32';
 const _SIGINT = 'SIGINT';
@@ -26,25 +26,34 @@ type SupportedPlatforms = 'darwin'|'linux'|'win32'|'wsl';
 
 const instances = new Set<Launcher>();
 
+type JSONLike =|{[property: string]: JSONLike}|readonly JSONLike[]|string|number|boolean|null;
+
 export interface Options {
   startingUrl?: string;
   edgeFlags?: Array<string>;
+  prefs?: Record<string, JSONLike>;
   port?: number;
+  portStrictMode?: boolean;
   handleSIGINT?: boolean;
   edgePath?: string;
   userDataDir?: string|boolean;
-  logLevel?: 'verbose'|'info'|'error'|'silent';
+  logLevel?: 'verbose'|'info'|'error'|'warn'|'silent';
   ignoreDefaultFlags?: boolean;
   connectionPollInterval?: number;
   maxConnectionRetries?: number;
   envVars?: {[key: string]: string|undefined};
 }
 
+export interface RemoteDebuggingPipes {
+  incoming: NodeJS.ReadableStream, outgoing: NodeJS.WritableStream,
+}
+
 export interface LaunchedEdge {
   pid: number;
   port: number;
   process: ChildProcess;
-  kill: () => Promise<void>;
+  remoteDebuggingPipes: RemoteDebuggingPipes|null;
+  kill: () => void;
 }
 
 export interface ModuleOverrides {
@@ -52,8 +61,8 @@ export interface ModuleOverrides {
   spawn?: typeof childProcess.spawn;
 }
 
-const sigintListener = async () => {
-  await killAll();
+const sigintListener = () => {
+  killAll();
   process.exit(_SIGINT_EXIT_CODE);
 };
 
@@ -70,22 +79,37 @@ async function launch(opts: Options = {}): Promise<LaunchedEdge> {
 
   await instance.launch();
 
-  const kill = async () => {
+  const kill = () => {
     instances.delete(instance);
     if (instances.size === 0) {
       process.removeListener(_SIGINT, sigintListener);
     }
-    return instance.kill();
+    instance.kill();
   };
 
-  return {pid: instance.pid!, port: instance.port!, kill, process: instance.edge!};
+  return {
+    pid: instance.pid!,
+    port: instance.port!,
+    process: instance.edgeProcess!,
+    remoteDebuggingPipes: instance.remoteDebuggingPipes,
+    kill,
+  };
 }
 
-async function killAll(): Promise<Array<Error>> {
+/** Returns Edge installation path that chromium-edge-launcher will launch by default. */
+function getEdgePath(): string {
+  const installation = Launcher.getFirstInstallation();
+  if (!installation) {
+    throw new EdgeNotInstalledError();
+  }
+  return installation;
+}
+
+function killAll(): Array<Error> {
   let errors = [];
   for (const instance of instances) {
     try {
-      await instance.kill();
+      instance.kill();
       // only delete if kill did not error
       // this means erroring instances remain in the Set
       instances.delete(instance);
@@ -105,7 +129,10 @@ class Launcher {
   private edgePath?: string;
   private ignoreDefaultFlags?: boolean;
   private edgeFlags: string[];
+  private prefs: Record<string, JSONLike>;
   private requestedPort?: number;
+  private portStrictMode?: boolean;
+  private useRemoteDebuggingPipe: boolean;
   private connectionPollInterval: number;
   private maxConnectionRetries: number;
   private fs: typeof fs;
@@ -113,9 +140,10 @@ class Launcher {
   private useDefaultProfile: boolean;
   private envVars: {[key: string]: string|undefined};
 
-  edge?: childProcess.ChildProcess;
+  edgeProcess?: childProcess.ChildProcess;
   userDataDir?: string;
   port?: number;
+  remoteDebuggingPipes: RemoteDebuggingPipes|null = null;
   pid?: number;
 
   constructor(private opts: Options = {}, moduleOverrides: ModuleOverrides = {}) {
@@ -127,7 +155,9 @@ class Launcher {
     // choose the first one (default)
     this.startingUrl = defaults(this.opts.startingUrl, 'about:blank');
     this.edgeFlags = defaults(this.opts.edgeFlags, []);
+    this.prefs = defaults(this.opts.prefs, {});
     this.requestedPort = defaults(this.opts.port, 0);
+    this.portStrictMode = opts.portStrictMode;
     this.edgePath = this.opts.edgePath;
     this.ignoreDefaultFlags = defaults(this.opts.ignoreDefaultFlags, false);
     this.connectionPollInterval = defaults(this.opts.connectionPollInterval, 500);
@@ -145,11 +175,18 @@ class Launcher {
       this.useDefaultProfile = false;
       this.userDataDir = this.opts.userDataDir;
     }
+
+    // Using startsWith because it could also be --remote-debugging-pipe=cbor
+    this.useRemoteDebuggingPipe =
+        this.edgeFlags.some(f => f.startsWith('--remote-debugging-pipe'));
   }
 
   private get flags() {
     const flags = this.ignoreDefaultFlags ? [] : DEFAULT_FLAGS.slice();
-    flags.push(`--remote-debugging-port=${this.port}`);
+    // When useRemoteDebuggingPipe is true, this.port defaults to 0.
+    if (this.port) {
+      flags.push(`--remote-debugging-port=${this.port}`);
+    }
 
     if (!this.ignoreDefaultFlags && getPlatform() === 'linux') {
       flags.push('--disable-setuid-sandbox');
@@ -158,8 +195,10 @@ class Launcher {
     if (!this.useDefaultProfile) {
       // Place Edge profile in a custom location we'll rm -rf later
       // If in WSL, we need to use the Windows format
-      flags.push(`--user-data-dir=${isWsl ? toWinDirFormat(this.userDataDir) : this.userDataDir}`);
+      flags.push(`--user-data-dir=${isWsl ? toWin32Path(this.userDataDir) : this.userDataDir}`);
     }
+
+    if (process.env.HEADLESS) flags.push('--headless');
 
     flags.push(...this.edgeFlags);
     flags.push(this.startingUrl);
@@ -177,6 +216,11 @@ class Launcher {
     return edgeFinder[getPlatform() as SupportedPlatforms]()[0];
   }
 
+  /** Returns all available edge installations in decreasing priority order. */
+  static getInstallations() {
+    return edgeFinder[getPlatform() as SupportedPlatforms]();
+  }
+
   // Wrapper function to enable easy testing.
   makeTmpDir() {
     return makeTmpDir();
@@ -192,6 +236,8 @@ class Launcher {
     this.outFile = this.fs.openSync(`${this.userDataDir}/edge-out.log`, 'a');
     this.errFile = this.fs.openSync(`${this.userDataDir}/edge-err.log`, 'a');
 
+    this.setBrowserPrefs();
+
     // fix for Node4
     // you can't pass a fd to fs.writeFileSync
     this.pidFile = `${this.userDataDir}/edge.pid`;
@@ -201,16 +247,52 @@ class Launcher {
     this.tmpDirandPidFileReady = true;
   }
 
+  private setBrowserPrefs() {
+    // don't set prefs if not defined
+    if (Object.keys(this.prefs).length === 0) {
+      return;
+    }
+
+    const profileDir = `${this.userDataDir}/Default`;
+    if (!this.fs.existsSync(profileDir)) {
+      this.fs.mkdirSync(profileDir, {recursive: true});
+    }
+
+    const preferenceFile = `${profileDir}/Preferences`;
+    try {
+      if (this.fs.existsSync(preferenceFile)) {
+        // overwrite existing file
+        const file = this.fs.readFileSync(preferenceFile, 'utf-8');
+        const content = JSON.parse(file);
+        this.fs.writeFileSync(preferenceFile, JSON.stringify({...content, ...this.prefs}), 'utf-8');
+      } else {
+        // create new Preference file
+        this.fs.writeFileSync(preferenceFile, JSON.stringify({...this.prefs}), 'utf-8');
+      }
+    } catch (err) {
+      log.log('EdgeLauncher', `Failed to set browser prefs: ${err.message}`);
+    }
+  }
+
   async launch() {
     if (this.requestedPort !== 0) {
       this.port = this.requestedPort;
 
       // If an explict port is passed first look for an open connection...
       try {
-        return await this.isDebuggerReady();
-      } catch (err) {
+        await this.isDebuggerReady();
         log.log(
-            'EdgeLauncher', `No debugging port found on port ${this.port}, launching a new Edge.`);
+            'EdgeLauncher',
+            `Found existing Edge already running using port ${this.port}, using that.`);
+        return;
+      } catch (err) {
+        if (this.portStrictMode) {
+          throw new Error(`found no Edge at port ${this.requestedPort}`);
+        }
+
+        log.log(
+            'EdgeLauncher',
+            `No debugging port found on port ${this.port}, launching a new Edge.`);
       }
     }
     if (this.edgePath === undefined) {
@@ -232,9 +314,9 @@ class Launcher {
 
   private async spawnProcess(execPath: string) {
     const spawnPromise = (async () => {
-      if (this.edge) {
-        log.log('EdgeLauncher', `Edge already running with pid ${this.edge.pid}.`);
-        return this.edge.pid;
+      if (this.edgeProcess) {
+        log.log('EdgeLauncher', `Edge already running with pid ${this.edgeProcess.pid}.`);
+        return this.edgeProcess.pid;
       }
 
 
@@ -243,23 +325,48 @@ class Launcher {
       // We do this here so that we can know the port before
       // we pass it into edge.
       if (this.requestedPort === 0) {
-        this.port = await getRandomPort();
+        if (this.useRemoteDebuggingPipe) {
+          // When useRemoteDebuggingPipe is true, this.port defaults to 0.
+          this.port = 0;
+        } else {
+          this.port = await getRandomPort();
+        }
       }
 
-      log.verbose('EdgeLauncher', `Launching with command:\n"${execPath}" ${this.flags.join(' ')}`);
-      const edge = this.spawn(
-          execPath, this.flags,
-          {detached: true, stdio: ['ignore', this.outFile, this.errFile], env: this.envVars});
-      this.edge = edge;
+      log.verbose(
+          'EdgeLauncher', `Launching with command:\n"${execPath}" ${this.flags.join(' ')}`);
+      this.edgeProcess = this.spawn(execPath, this.flags, {
+        // On non-windows platforms, `detached: true` makes child process a leader of a new
+        // process group, making it possible to kill child process tree with `.kill(-pid)` command.
+        // @see https://nodejs.org/api/child_process.html#child_process_options_detached
+        detached: process.platform !== 'win32',
+        stdio: this.useRemoteDebuggingPipe ?
+            ['ignore', this.outFile, this.errFile, 'pipe', 'pipe'] :
+            ['ignore', this.outFile, this.errFile],
+        env: this.envVars
+      });
 
-      this.fs.writeFileSync(this.pidFile, edge.pid!.toString());
+      if (this.edgeProcess.pid) {
+        this.fs.writeFileSync(this.pidFile, this.edgeProcess.pid.toString());
+      }
+      if (this.useRemoteDebuggingPipe) {
+        this.remoteDebuggingPipes = {
+          incoming: this.edgeProcess.stdio[4] as NodeJS.ReadableStream,
+          outgoing: this.edgeProcess.stdio[3] as NodeJS.WritableStream,
+        };
+      }
 
-      log.verbose('EdgeLauncher', `Edge running with pid ${edge.pid} on port ${this.port}.`);
-      return edge.pid;
+      log.verbose(
+          'EdgeLauncher',
+          `Edge running with pid ${this.edgeProcess.pid} on port ${this.port}.`);
+      return this.edgeProcess.pid;
     })();
 
     const pid = await spawnPromise;
-    await this.waitUntilReady();
+    // When useRemoteDebuggingPipe is true, this.port defaults to 0.
+    if (this.port !== 0) {
+      await this.waitUntilReady();
+    }
     return pid;
   }
 
@@ -275,7 +382,11 @@ class Launcher {
   // resolves if ready, rejects otherwise
   private isDebuggerReady(): Promise<void> {
     return new Promise((resolve, reject) => {
-      const client = net.createConnection(this.port!);
+      // Note: only meaningful when this.port is set.
+      // When useRemoteDebuggingPipe is true, this.port defaults to 0. In that
+      // case, we could consider ping-ponging over the pipe, but that may get
+      // in the way of the library user, so we do not.
+      const client = net.createConnection(this.port!, '127.0.0.1');
       client.once('error', err => {
         this.cleanup(client);
         reject(err);
@@ -313,7 +424,8 @@ class Launcher {
                 log.error('EdgeLauncher', err.message);
                 const stderr =
                     this.fs.readFileSync(`${this.userDataDir}/edge-err.log`, {encoding: 'utf-8'});
-                log.error('EdgeLauncher', `Logging contents of ${this.userDataDir}/edge-err.log`);
+                log.error(
+                    'EdgeLauncher', `Logging contents of ${this.userDataDir}/edge-err.log`);
                 log.error('EdgeLauncher', stderr);
                 return reject(err);
               }
@@ -325,55 +437,58 @@ class Launcher {
   }
 
   kill() {
-    return new Promise<void>((resolve, reject) => {
-      if (this.edge) {
-        this.edge.on('close', () => {
-          delete this.edge;
-          this.destroyTmp().then(resolve);
-        });
+    if (!this.edgeProcess) {
+      return;
+    }
 
-        log.log('EdgeLauncher', `Killing Edge instance ${this.edge.pid}`);
-        try {
-          if (isWindows) {
-            // While pipe is the default, stderr also gets printed to process.stderr
-            // if you don't explicitly set `stdio`
-            execSync(`taskkill /pid ${this.edge.pid} /T /F`, {stdio: 'pipe'});
-          } else {
-            process.kill(-this.edge.pid!);
-          }
-        } catch (err) {
-          const message = `Edge could not be killed ${err.message}`;
-          log.warn('EdgeLauncher', message);
-          reject(new Error(message));
-        }
-      } else {
-        // fail silently as we did not start edge
-        resolve();
-      }
+    this.edgeProcess.on('close', () => {
+      delete this.edgeProcess;
+      this.destroyTmp();
     });
+
+    log.log('EdgeLauncher', `Killing Edge instance ${this.edgeProcess.pid}`);
+    try {
+      if (isWindows) {
+        // https://github.com/GoogleChrome/chrome-launcher/issues/266
+        const taskkillProc = spawnSync(
+            `taskkill /pid ${this.edgeProcess.pid} /T /F`, {shell: true, encoding: 'utf-8'});
+
+        const {stderr} = taskkillProc;
+        if (stderr) log.error('EdgeLauncher', `taskkill stderr`, stderr);
+      } else {
+        if (this.edgeProcess.pid) {
+          process.kill(-this.edgeProcess.pid, 'SIGKILL');
+        }
+      }
+    } catch (err) {
+      const message = `Edge could not be killed ${err.message}`;
+      log.warn('EdgeLauncher', message);
+    }
+    this.destroyTmp();
   }
 
   destroyTmp() {
-    return new Promise<void>(resolve => {
-      // Only clean up the tmp dir if we created it.
-      if (this.userDataDir === undefined || this.opts.userDataDir !== undefined) {
-        return resolve();
-      }
+    if (this.outFile) {
+      this.fs.closeSync(this.outFile);
+      delete this.outFile;
+    }
 
-      if (this.outFile) {
-        this.fs.closeSync(this.outFile);
-        delete this.outFile;
-      }
+    // Only clean up the tmp dir if we created it.
+    if (this.userDataDir === undefined || this.opts.userDataDir !== undefined) {
+      return;
+    }
 
-      if (this.errFile) {
-        this.fs.closeSync(this.errFile);
-        delete this.errFile;
-      }
+    if (this.errFile) {
+      this.fs.closeSync(this.errFile);
+      delete this.errFile;
+    }
 
-      this.fs.rmdir(this.userDataDir, {recursive: true}, () => resolve());
-    });
+    // backwards support for node v12 + v14.14+
+    // https://nodejs.org/api/deprecations.html#DEP0147
+    const rmSync = this.fs.rmSync || this.fs.rmdirSync;
+    rmSync(this.userDataDir, {recursive: true, force: true, maxRetries: 10});
   }
 };
 
 export default Launcher;
-export {Launcher, launch, killAll};
+export {Launcher, launch, killAll, getEdgePath};