+++ test/launch-signature-test.ts
@@ -6,11 +6,11 @@
 
 'use strict';
 
-// @ts-ignore
-import {launch} from '../src/';
+import {launch} from '../src/index.js';
 import * as assert from 'assert';
 
-const log = require('lighthouse-logger');
+import log from 'lighthouse-logger';
+
 describe('Launcher', () => {
 
   beforeEach(() => {
@@ -27,6 +27,6 @@ describe('Launcher', () => {
     assert.notStrictEqual(edge.pid, undefined);
     assert.notStrictEqual(edge.port, undefined);
     assert.notStrictEqual(edge.kill, undefined);
-    await edge.kill();
+    edge.kill();
   });
 });