+++ changelog.md
@@ -1,3 +1,66 @@
+## v1.0.0 (Tue, Jul 18 2023)
+
+* `37609a01` deps: upgrade `lighthouse-logger` to 2.0.1 ([#309](https://github.com/GoogleChrome/chrome-launcher/pull/309))
+* `dacd3578` Publish as ESM ([#302](https://github.com/GoogleChrome/chrome-launcher/pull/302))
+* `f9e43693` fix: always close file handles to stdout/stderr logs ([#259](https://github.com/GoogleChrome/chrome-launcher/pull/259))
+
+## v0.15.2 (Mon, April 17 2023)
+
+* `76b6a13c` Update logLevel option typing ([#295](https://github.com/GoogleChrome/chrome-launcher/pull/295))
+* `60044483` Update headless=chrome flag to headless=new ([#290](https://github.com/GoogleChrome/chrome-launcher/pull/290))
+* `b041125a` log when connection to existing chrome found for requested port ([#291](https://github.com/GoogleChrome/chrome-launcher/pull/291))
+* `f64a7d89` docs(flags): fix description of mock-keychain flag
+* `c753ba08` use updating node versions in CI ([#286](https://github.com/GoogleChrome/chrome-launcher/pull/286))
+* `8d1d85dc` docs(flags): adjust grouping. add in several features ([#283](https://github.com/GoogleChrome/chrome-launcher/pull/283))
+* `471a97c7` flags: disable optimization guide and media router ([#282](https://github.com/GoogleChrome/chrome-launcher/pull/282))
+* `83f08461` docs(flags): add disable-features=MediaRouter which is surprisingly active
+* `346b3c2c` rename master branch references to main ([#280](https://github.com/GoogleChrome/chrome-launcher/pull/280))
+* `f618e7e5` docs: a few additions to the flags doc
+* `c36bd8dc` docs: add OptimizationHints to flags
+* `1cbf8b9a` Make LaunchedChrome.kill sync ([#269](https://github.com/GoogleChrome/chrome-launcher/pull/269))
+
+
+## v0.15.1 (Tue, May 31 2022)
+
+* `3724165a` make launcher.kill() synchronous ([#268](https://github.com/GoogleChrome/chrome-launcher/pull/268))
+* `3561350a` revise taskkill procedure on windows ([#267](https://github.com/GoogleChrome/chrome-launcher/pull/267))
+* `690ae983` add lighthouse smoketests (windows) to CI ([#265](https://github.com/GoogleChrome/chrome-launcher/pull/265))
+* `279577fd` docs(edge-flags-for-tools): add link to overview of features ([#235](https://github.com/GoogleChrome/chrome-launcher/pull/235))
+* `ff91c18b` fix: use `wslpath` to resolve Windows paths ([#200](https://github.com/GoogleChrome/chrome-launcher/pull/200))
+* `30755cde` test: run latest versions of node in CI ([#257](https://github.com/GoogleChrome/chrome-launcher/pull/257))
+
+## v0.15.0 (Wed, Nov 10 2021)
+
+* `1af60cb` add `getEdgePath()` method for printing found chrome path ([#255](https://github.com/GoogleChrome/chrome-launcher/pull/255))
+
+## v0.14.2 (Tue, Nov 2 2021)
+
+* `ba8d76bd` fix chrome connection in node 17, use 127.0.0.1 explicitly ([#253](https://github.com/GoogleChrome/chrome-launcher/pull/253))
+* `56731dd8` fix: forward support for fs.rm in node 14.14+ ([#242](https://github.com/GoogleChrome/chrome-launcher/pull/242))
+
+## v0.14.1 (Wed, Oct 6 2021)
+
+* `630bb77f` feat: set browser preferences ([#247](https://github.com/GoogleChrome/chrome-launcher/pull/247))
+* `12b2c8e3` docs(flags): note that disable-gpu isnt needed
+
+
+## v0.14.0 (Tue, May 18 2021)
+* `ac1f4aff` move to minimum node 12; remove `rimraf` ([#237](https://github.com/GoogleChrome/chrome-launcher/pull/237))
+* `dec646c4` deps: remove `mkdirp` for `fs.mkdirSync` ([#234](https://github.com/GoogleChrome/chrome-launcher/pull/234))
+* `83ab178a` update minimum node version ([#222](https://github.com/GoogleChrome/chrome-launcher/pull/222))
+* `a5f6eb2f` add additional chrome flags ([#227](https://github.com/GoogleChrome/chrome-launcher/pull/227))
+* `3a7c9610` reword unset-`EDGE_PATH` error message
+* `b1b8dc74` rename disabled `TranslateUI` to `Translate` to match Chrome ([#225](https://github.com/GoogleChrome/chrome-launcher/pull/225))
+* `beb41360` chore: update dependencies and test targets ([#221](https://github.com/GoogleChrome/chrome-launcher/pull/221))
+* `df9d564a` tests: migrate from travis to github actions ([#228](https://github.com/GoogleChrome/chrome-launcher/pull/228))
+* `673da08b` tests: add mac/win bots to ci ([#232](https://github.com/GoogleChrome/chrome-launcher/pull/232))
+* `a700ae0c` docs: fix readme's `getInstallations()` section ([#212](https://github.com/GoogleChrome/chrome-launcher/pull/212))
+* [`edge-flags-for-tools.md`](https://github.com/GoogleChrome/chrome-launcher/blob/b00fa22f94371f6d38d2d23e6e0b32fc7af470f0/docs/chrome-flags-for-tools.md) update
+  - `4b98587d` massive update and refactor ([#226](https://github.com/GoogleChrome/chrome-launcher/pull/226))
+  - `e45b100f` minor tweaks to headless and others
+  - `3a90c21b` fix links
+  - `8429ca93` add feature flags description
+  - `21db5f9f` even more documented flags ([#231](https://github.com/GoogleChrome/chrome-launcher/pull/231))
 
 ## v0.13.4 (Tue, Jul 7 2020)
 * `08406b28` fix: preserve existing getInstallations output
@@ -61,8 +124,8 @@
 * `179a3f33` silence grep (#138)
 * `d2f6037a` fix: move unneeded ts types to devDeps (#142)
 * `984d61ce` docs(flags): remove a few flags that are gone.
-* `6316362c` docs: fix link to chrome-launcher's flags (#128)
-* `f1f6d162` Update chrome-flags-for-tools.md
+* `6316362c` docs: fix link to chromium-edge-launcher's flags (#128)
+* `f1f6d162` Update edge-flags-for-tools.md
 
 ## v0.10.5 (Tue, Sep 25 2018)
 * `1328319b` fix: set the `which` command's stdio to pipe (#125)
@@ -71,13 +134,13 @@
 * `35842ba4` fix: ignore stdio on `which` call (#121)
 * `f126c3a0` fix: reject promise on failed kill() (#112)
 * `5ee0fde2` Set custom error codes for all errors.
-* `841bdf3f` Fix picking CHROME_PATH priority over other matches.
+* `841bdf3f` Fix picking EDGE_PATH priority over other matches.
 * `6b10d748` Fix Travis CI build: GCE for chrome bug (#87)
 * `d4aa8295` Fix readme's default logLevel (#85)
 * `5be71243` Type improvements (#102)
 * `dd5fdd49` Stricter typing for logLevel (#105)
 * `c9394cf7` Fix README typo: booelan ==> boolean (#104)
-* Update chrome-flags-for-tools.md
+* Update edge-flags-for-tools.md
 
 ## v0.10.3 (Mon, Sep 17 2018)
 Bad release. Had a breaking change (#70). Unpublished.
@@ -88,7 +151,7 @@ Bad release. Had a breaking change (#70). Unpublished.
 
 ## v0.10.1 (Fri, Jan 5 2018)
 * `a5bc8180` Fix getLocalAppDataPath for wsl (#75)
-* `70a91885` readme: recommend use of cri with chrome-launcher (#78)
+* `70a91885` readme: recommend use of cri with chromium-edge-launcher (#78)
 * `d3ee63bd` folder refactor: ts in /src, js in /dist (#69)
 
 ## 0.10.0 (Fri, Dec 8 2017)
@@ -120,26 +183,26 @@ Bad release. Had a breaking change (#70). Unpublished.
 * `3ab3a117` docs: add changelog to launcher (#2987)
 
 ## 0.5.0 (Mon, 14 Aug 2017)
-* `494f9911` clarify priority of chromePath options
+* `494f9911` clarify priority of edgePath options
 * `1c11021a` add support for finding Chromium on Linux (#2950)
 * `391e2043` Publish type definitions instead of source TypeScript files (#2898)
 * `de408ad3` readme: update example using deprecated `LIGHTHOUSE_CHROMIUM_PATH` (#2929)
 * `8bc6d18e` add license file to launcher package. (#2849)
 
 ## 0.4.0 (Tue, 1 Aug 2017)
-* `37fd38ce` pass --enable-extensions on from manual-chrome-launcher (#2735)
+* `37fd38ce` pass --enable-extensions on from manual-edge-launcher (#2735)
 * `c942d17e` support enabling extension loading (#2650)
 
 ## 0.3.2 (Wed, 19 Jul 2017)
-* `112c2c7f` Fix chrome finder on linux/osx when process.env isn't populated (#2687)
-* `5728695f` Added CHROME_PATH to readme (#2694)
+* `112c2c7f` Fix edge finder on linux/osx when process.env isn't populated (#2687)
+* `5728695f` Added EDGE_PATH to readme (#2694)
 * `fedc76a3` test: fix clang-format error (#2691)
 * `a6bbcaba` nuke 'as string'
-* `41df647f` cli: remove --select-chrome,--skip-autolaunch. Support CHROME_PATH env  (#2659)
+* `41df647f` cli: remove --select-edge,--skip-autolaunch. Support EDGE_PATH env  (#2659)
 * `8c9724e2` fix launcher w/ arbitrary flags (#2670)
 * `9c0c0788` Expose LHR to modules consuming cli/run.ts (#2654)
-* `6df6b0e2` support custom port via chrome-debug binary (#2644)
-* `3f143b19` log the specific chrome spawn command.
+* `6df6b0e2` support custom port via edge-debug binary (#2644)
+* `3f143b19` log the specific edge spawn command.
 
 ## 0.3.1 (Wed, 5 Jul 2017)
 * `ef081063` upgrade rimraf to latest (#2641)